// collaboration-service/proto/generatedTypes/collaboration/ConnectionResponse.ts
// Original file: proto/collaboration.proto

import type { CollaborationAction as _collaboration_CollaborationAction, CollaborationAction__Output as _collaboration_CollaborationAction__Output } from '../collaboration/CollaborationAction';

export interface ConnectionResponse {
  'action'?: (_collaboration_CollaborationAction);
  'data'?: (string);
}

export interface ConnectionResponse__Output {
  'action'?: (_collaboration_CollaborationAction__Output);
  'data'?: (string);
}